this is the basic version of blog chap 1-25, without class views. please use article/oldviews.py instead of views.py

+++ article/views.py
@@ -1,5 +1,5 @@
-from django.shortcuts import render
-
+#基于类视图
+# from django.shortcuts import render
 # Create your views here.
 from django.http import HttpResponse
 from django.shortcuts import render
@@ -12,44 +12,55 @@ from .forms import ArticlePostForm
 from django.contrib.auth.decorators import login_required
 # 引入User模型
 from django.contrib.auth.models import User
+# 引入分页模块
+from django.core.paginator import Paginator
+# 引入 Q 对象
+from django.db.models import Q
+from comment.models import Comment
 
-def article_list(request):
-    #ArticlePost.objects.all()是数据类的方法，可以获得所有的对象（即博客文章），并传递给articles变量
-    articles = ArticlePost.objects.all()
-    #context定义了需要传递给模板的上下文，这里即articles
-    context = {'articles': articles}
-    #render函数的作用是结合模板和上下文，并返回渲染后的HttpResponse对象。
-    #通俗的讲就是把context的内容，加载进模板，并通过浏览器呈现。
-    # equest是固定的request对象，照着写就可以
-    # article / list.html定义了模板文件的位置、名称
-    # context定义了需要传入模板文件的上下文
-    return render(request, 'article/list.html', context)
+from django.views.generic import ListView
+from django.views.generic import DetailView
+from django.views.generic import CreateView
+class ContextMixin:
+    '''
+    混入类（Mixin）是指具有某些功能、通常不独立使用、提供给其他类继承功能的类。嗯，就是“混入”的字面意思。
+    '''
+    def get_context_data(self, **kwargs):
+        # 获取原有的上下文
+        context = super().get_context_data(**kwargs)
+        # 增加新上下文
+        context['order'] = 'total_views'
+        return context
 
-#article_detail(request, id)函数中多了id这个参数。**注意我们在写model的时候并没有写叫做id的字段，
-# **这是Django自动生成的用于索引数据表的主键（Primary Key，即pk）。
-# 有了它才有办法知道到底应该取出哪篇文章。
-def article_detail(request, id):
-    #意思是在所有文章中，取出id值相符合的唯一的一篇文章。
-    article = ArticlePost.objects.get(id = id)
-    # 将markdown语法渲染成html样式,将Markdown语法书写的文章渲染为HTML文本
-    article.body = markdown.markdown(article.body,
-                                     extensions=[
-                                         # 包含 缩写、表格等常用扩展
-                                         'markdown.extensions.extra',
-                                         # 语法高亮扩展
-                                         'markdown.extensions.codehilite',
-                                     ])
-    context = {'article':article}
-    return render(request, 'article/detail.html', context)
+class ArticleListView(ContextMixin, ListView):
+    #通过混入，两个子类都获得了get_context_data()方法。
+    #从语法上看，混入是通过多重继承实现的。有区别的是，Mixin是作为功能添加到子类中的，而不是作为父类。
+    #Django内置了很多通用的Mixin类，实现了大部分常用的功能，可以往官方文档进一步了解
+    context_object_name = 'articles'
+    template_name = 'article/list.html'
+    def get_queryset(self):
+        """
+        查询集
+        """
+        queryset = ArticlePost.objects.filter(title='Python')
+        return queryset
 
-def article_delete(request, id):
-    # 意思是在所有文章中，取出id值相符合的唯一的一篇文章。
-    article = ArticlePost.objects.get(id=id)
-    # 调用.delete()方法删除文章
-    article.delete()
-    # 完成删除后返回文章列表,第一个article是app名字
-    # 第二个article_list是在urls里面我们定义的名字
-    return redirect("article:article_list")
+
+class ArticleDetailView(DetailView):
+    queryset = ArticlePost.objects.all()
+    context_object_name = 'article'
+    template_name = 'article/detail.html'
+    def get_object(self):
+        """
+        获取需要展示的对象
+        """
+        # 首先调用父类的方法
+        obj = super(ArticleDetailView, self).get_object()
+        # 浏览量 +1
+        obj.total_views += 1
+        obj.save(update_fields=['total_views'])
+        return obj
+
 
 def article_safe_delete(request, id):
     '''
@@ -65,6 +76,8 @@ def article_safe_delete(request, id):
     if request.method == 'POST':
         # 意思是在所有文章中，取出id值相符合的唯一的一篇文章。
         article = ArticlePost.objects.get(id=id)
+        if request.user != article.author:
+            return HttpResponse('抱歉，你无权删除这篇文章')
         # 调用.delete()方法删除文章
         article.delete()
         # 完成删除后返回文章列表,第一个article是app名字
@@ -72,45 +85,11 @@ def article_safe_delete(request, id):
         return redirect("article:article_list")
     else:
         return HttpResponse('POST request only!')
-'''
-#profile修改之前的article_create:
-#两点问题:  1. new_article.author = User.objects.get(id=1)强行把作者指定为id=1的用户，这显然是不对的。
-          2. 没有对用户的登录状态进行检查。
-#解决:     1. User.objects.get(id=request.user.id)
-          2. 添加装饰器login_required
-def article_create(request):
-    if request.method == 'POST':
-        article_post_form = ArticlePostForm(data=request.POST)
-        if article_post_form.is_valid():
-            new_article = article_post_form.save(commit=False)
-            new_article.author = User.objects.get(id=1)
-            new_article.save()
-            return redirect("article:article_list")
-        else:
-            return HttpResponse('invalid form, please imput again!')
-    else:
-        article_post_form = ArticlePostForm()
-        context = {'article_post_form':article_post_form}
-        return render(request, 'article/create.html', context)
-'''
 
-@login_required(login_url='/userprofile/login/')
-def article_create(request):
-
-    if request.method == 'POST':
-        article_post_form = ArticlePostForm(data=request.POST)
-        if article_post_form.is_valid():
-            new_article = article_post_form.save(commit=False)
-            #new_article.author = User.objects.get(id=1)
-            new_article.author = User.objects.get(id=request.user.id)
-            new_article.save()
-            return redirect("article:article_list")
-        else:
-            return HttpResponse('invalid form, please imput again!')
-    else:
-        article_post_form = ArticlePostForm()
-        context = {'article_post_form': article_post_form}
-        return render(request, 'article/create.html', context)
+class ArticleCreateView(CreateView):
+    model = ArticlePost
+    fields = '__all__'
+    template_name = 'article/create_by_class_view.html'
 
 def article_update(request, id):
     """
@@ -120,6 +99,8 @@ def article_update(request, id):
         id： 文章的 id
     """
     article = ArticlePost.objects.get(id=id)
+    if request.user != article.author:
+        return HttpResponse('抱歉，你无权修改这篇文章')
     if request.method == 'POST':
         article_post_form = ArticlePostForm(data=request.POST)
         if article_post_form.is_valid():
